add handling for same departure and destination
