+++ src/main/java/com/mitrais/rnd/FlightManagementSystem/constant/ErrorMesssageConstant.java
@@ -2,4 +2,5 @@ package com.mitrais.rnd.FlightManagementSystem.constant;
 
 public class ErrorMesssageConstant {
     public static final String DESTINATION_ALREADY_EXISTS  = "Destination name already exists please use other name";
+    public static final String SAME_DEPARTURE_DESTINATION_ERROR  = "Departure and destination must be different. Please check your input.";
 }
