+++ src/main/java/com/mitrais/rnd/FlightManagementSystem/service/RouteService.java
@@ -4,5 +4,5 @@ import com.mitrais.rnd.FlightManagementSystem.entity.Route;
 import jakarta.persistence.EntityExistsException;
 
 public interface RouteService {
-    public void addRoute(Route route);
+    public void addRoute(Route route) throws Exception;
 }
